+++ app/views/schedules/_calendar.html.erb
@@ -1,17 +1,19 @@
-<%= month_calendar events: @schedules, attribute: :date do |date, schedules| %>
-  <div class="day calendar-day position-relative">
-    <%= date.day %>
-    <% if schedules.count > 0 %>
-      <span class="badge text-bg-cYellow rounded-circle position-absolute calendar-badge" style="top: 3px; right: 0;">
-        <%= schedules.count %>
-      </span>
-    <% end %>
-    <div class="calendar-event">
-      <% schedules.each do |schedule| %>
-        <div class="event mb-1">
-          <%= link_to schedule.recipe.name, recipe_path(schedule.recipe),class: "text-decoration-none text-cBlack" %>
-        </div>
+<div id="calendar" style="overflow-x:auto;max-width:400px">
+  <%= month_calendar events: schedules, attribute: :date do |date, schedules_day| %>
+    <div class="day calendar-day position-relative">
+      <%= date.day %>
+      <% if schedules_day.count > 0 %>
+        <span class="badge text-bg-cYellow rounded-circle position-absolute calendar-badge" style="top: 3px; right: 0;">
+          <%= schedules_day.count %>
+        </span>
       <% end %>
+      <div class="calendar-event">
+        <% schedules_day.each do |schedule| %>
+          <div class="event mb-1">
+            <%= link_to schedule.recipe.name, recipe_path(schedule.recipe),class: "text-decoration-none text-cBlack" %>
+          </div>
+        <% end %>
+      </div>
     </div>
-  </div>
-<% end %>
+  <% end %>
+</div>
