+++ app/views/schedules/generate_ai.turbo_stream.erb
@@ -1,6 +1,6 @@
 <%= turbo_stream.replace "divFormAi" do %>
-  <%= render "schedules/ai", schedule: Schedule.new %>
+  <%= render "schedules/ai", schedule: Schedule.new, start_date: @start_date %>
 <% end %>
-<%= turbo_stream.update "calendar" do %>
+<%= turbo_stream.replace "calendar" do %>
   <%= render "schedules/calendar", schedules: @schedules %>
 <% end %>
