+++ app/views/schedules/index.html.erb
@@ -1,4 +1,5 @@
-<div class="container" data-controller="scheduler">
+<%= turbo_stream_from "schedules" %>
+<div class="container" data-controller="scheduler load">
   <h2 class="text-cWhite mt-5 mb-4">Your Meal Plan</h2>
   <%= link_to "Add to Meal Plan", "#",
     class: "btn btn-cDarkYellow mb-3 rounded-4",
@@ -49,8 +50,6 @@
         <i class="fa-solid fa-arrow-right"></i>
       <% end %>
     </div>
-    <div id="calendar" style="overflow-x:auto;max-width:400px">
-      <%= render "schedules/calendar", schedules: @schedules %>
-    </div>
+    <%= render "schedules/calendar", schedules: @schedules %>
   </div>
 </div>
