+++ test/jobs/schedulerbroadcast_job_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class SchedulerbroadcastJobTest < ActiveJob::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
